:wrench: Add a requirements.txt file for the project
The requirements.txt file is a text file separated by line feeds
listing the third-party modules used in the project. This file
can be supplied to the 'pip' CLI tool to quickly install all
the modules:
    pip install -r requirements.txt

It is recommended to sandbox each project using Python virtual
environments, instead of installing modules globally in your
system. To create a new virtual environment in the current directory:
    python -m venv .env

This will create a folder named '.env' for a Python virtual
environment of the same name. After this, you can activate it. See
Python's virtual environment official documentation to learn more).